requirements.txt
python-dateutil
pytest
email-validator
mock
psycopg2